Add installer stage methods
