--- install/lib/installer_stage_2.php
@@ -1,3 +1,0 @@
-<?php
-output("`#By continuing with this installation, you indicate your agreement with the terms of the license found on the previous page (License Agreement).");
-?>
